generated/requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
openai==0.27.0
gunicorn==20.1.0